requirements.txt
Flask
Flask-Login
Flask-WTF
email-validator
Pillow
SQLAlchemy
gunicorn